requirements.txt
celery[redis]
pyoai
PyPDF2
bibtexparser>=1.1.0
redis
cryptography
django>=2.2.0,<2.3.0
django-allauth
django-bootstrap-pagination
django-bulk-update
django-bootstrap-datepicker-plus
django-cache-machine
django-capture-tag
django-cas-ng
django-countries
django-crispy-forms
django-geojson
django-haystack
django-js-reverse
django-jsonview
django-leaflet
django-positions
django-memoize
django-redis-cache
django-settings-export
django-solo
django-ratelimit
django-widget-tweaks
django-debug-toolbar
django-vinaigrette
djangorestframework
elasticsearch==6.3.1
httplib2
jsonfield
langdetect
lxml
mock
multiprocessing_generator
name_tools
ndg-httpsclient
pillow
psycopg2-binary
pyasn1
pyopenssl>=0.14
python2-secrets ; python_version < '3.6'
reportlab
requests[security]
requests_mock
titlecase
unicode_tex
unidecode
urllib3
wand
sentry-sdk
python-dateutil # for haystack, as explained here:
#http://django-haystack.readthedocs.io/en/v2.4.1/management_commands.html
dealer
django-select2
bleach
